Start using mocha with istambul to test and setup circleci
