+++ circle.yml
@@ -1,0 +1,9 @@
+test:
+  override:
+    - nvm install 0.12 && npm test
+    - nvm install 4 && npm test
+  post:
+    - npm install
+    - istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec
+    - cat ./coverage/lcov.info | ./node_modules/.bin/codacy-coverage
+    - rm -rf ./coverage
